WIP: substantial refactoring, keep measurements outside of snapshots
Recording measurements inside the snapshots themselves makes the uuids
keep changing (changing a read-only snapshot involves copying/destroying
it). This makes life very difficult for snazzer-receive to do
incremental btrfs send -p updates.

Also tidied up escaping and stuff, but this is still a work in progress
so stuff may be broken and doc may be inconsistent with code... more to
come